fix: scanning errors

+++ infrastructure/cache/config.go
@@ -4,7 +4,7 @@ import "github.com/scrapnode/kanthor/pkg/validator"
 
 type Config struct {
 	Uri        string `json:"uri" yaml:"uri" mapstructure:"uri"`
-	TimeToLive int    `json:"time_to_live" yaml:"timeToLive" mapstructure:"time_to_live"`
+	TimeToLive uint64 `json:"time_to_live" yaml:"timeToLive" mapstructure:"time_to_live"`
 }
 
 func (conf *Config) Validate() error {
